% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqtag.R
\name{cumcount}
\alias{cumcount}
\title{Cumulative count of strings.}
\usage{
cumcount(xs)
}
\arguments{
\item{xs}{a character vector}
}
\value{
an integer vector of the cumulative string counts
}
\description{
Return an integer vector counting the number of occurrences of each string up to that position in the vector.
}
\examples{
cumcount(abbreviate(state.name, 3, strict = TRUE))
}
