% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqtag.R
\name{kmers_of}
\alias{kmers_of}
\alias{vkmers_of}
\title{Return the k-mers of a string.}
\usage{
kmers_of(x, k)

vkmers_of(xs, k)
}
\arguments{
\item{x}{a character string}

\item{k}{the size of the substrings, an integer}

\item{xs}{a character vector}
}
\value{
kmers_of: a character vector of the k-mers of \code{x}

vkmers_of: a list of character vectors of the k-mers of \code{xs}
}
\description{
Return the k-mers (substrings of size \code{k}) of the string \code{x}, or
return the string \code{x} itself if it is shorter than k.
}
\section{Functions}{
\itemize{
\item \code{kmers_of}: Return the k-mers of the string \code{x}.

\item \code{vkmers_of}: Return the k-mers of the strings \code{xs}.
}}

