% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqtag.R
\docType{package}
\name{uniqtag-package}
\alias{uniqtag-package}
\title{Abbreviate strings to short, unique identifiers.}
\description{
For each string in a set of strings, determine a unique tag that is a substring of fixed size k
unique to that string, if it has one. If no such unique substring exists, the least frequent
substring is used. If multiple unique substrings exist, the lexicographically smallest substring
is used. This lexicographically smallest substring of size k is called the "UniqTag" of that
string.
}
\author{
Shaun Jackman \email{sjackman@gmail.com}
}
