% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uniqtag.R
\docType{package}
\name{uniqtag-package}
\alias{uniqtag-package}
\title{Abbreviate strings to short, unique identifiers.}
\description{
Abbreviate strings to short, unique identifiers.
}
\author{
Shaun Jackman \email{sjackman@gmail.com}
}

